+++ crates/infra/audit/src/lib.rs
@@ -30,8 +30,8 @@ pub use rpc_reader::RpcEventReader;
 
 mod storage;
 pub use storage::{
-    BundleEventS3Reader, BundleHistory, BundleHistoryEvent, EventWriter, S3EventReaderWriter,
-    S3Key, TransactionMetadata,
+    BundleEventReader, BundleHistory, BundleHistoryEvent, EventWriter, RejectedTxStore,
+    S3EventReaderWriter, TransactionMetadata,
 };
 
 mod types;
